+++ packages/gateway/package.json
@@ -17,13 +17,11 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@react-portable/client": "0.5.0-next.3",
     "hono": "^3.3.0"
   },
   "devDependencies": {
     "@cloudflare/workers-types": "^4.20230710.0",
     "@vitest/coverage-v8": "^0.33.0",
-    "node-html-parser": "^6.1.5",
     "typescript": "^5.1.6",
     "vite": "^4.4.3",
     "vite-plugin-dts": "^3.2.0",
