+++ packages/gateway/src/index.ts
@@ -1,2 +1,1 @@
 export { gateway } from "./gateway";
-export { previewGateway } from "./preview-gateway";
