+++ packages/gateway/src/libs/htmlRewriters.test.ts
@@ -1,112 +1,10 @@
 import {
   ActivateRpPreviewReplacer,
   FragmentBaseReplacer,
-  FragmentTemplatesAppender,
-  ReactPortablePiercer,
 } from "./htmlRewriters";
-import { ReactPortable } from "@react-portable/client/web-components";
 import { beforeEach, describe, expect, test } from "vitest";
-import { FragmentMap } from "./fragments";
-
-const hostDummyResponseBody = `<!DOCTYPE html>
-<html>
-  <head>
-    <title>dummy page title</title>
-  </head>
-  <body>
-    <h1>dummy page</h1>
-    <react-portable entry="f1:/component1" gateway="https://gw1.com"></react-portable>
-    <react-portable entry="f2:/component2"></react-portable>
-    <react-portable entry="f3:/component3">has inner text</react-portable>
-    <react-portable>this is bad component (no entry attribute)</react-portable>
-    <react-portablee entry="f3:/component3">this is bad component (typo)</react-portablee>
-    <react-portable entry="f4:/component4">this is bad component (response status is 404)</react-portable>
-  </body>
-</html>
-`;
-
-const fragmentIdMapping: FragmentMap = new Map([
-  [
-    ReactPortable.createFragmentId("f1:/component1", "https://gw1.com"),
-    {
-      ok: true,
-      body: `<rp-fragment>this is component1</rp-fragment>`,
-      status: 200,
-      statusText: "OK",
-    },
-  ],
-  [
-    ReactPortable.createFragmentId("f2:/component2"),
-    {
-      ok: true,
-      body: `<rp-fragment>this is component2</rp-fragment>`,
-      status: 200,
-      statusText: "OK",
-    },
-  ],
-  [
-    ReactPortable.createFragmentId("f3:/component3"),
-    {
-      ok: true,
-      body: `<rp-fragment>this is component3</rp-fragment>`,
-      status: 200,
-      statusText: "OK",
-    },
-  ],
-  [
-    ReactPortable.createFragmentId("f4:/component4"),
-    {
-      ok: false,
-      body: null,
-      status: 404,
-      statusText: "Not Found",
-    },
-  ],
-]);
 
 describe("htmlRewriters", () => {
-  describe("ReactPortablePiercer", () => {
-    let response: Response;
-    beforeEach(() => {
-      response = new Response(hostDummyResponseBody);
-    });
-    test("empty fragmentIdsMap", async () => {
-      const piercer = new ReactPortablePiercer(new Map());
-      const rewriter = new HTMLRewriter().on(piercer.selector, piercer);
-
-      expect(await rewriter.transform(response).text()).toBe(
-        hostDummyResponseBody,
-      );
-    });
-    test("fulfilled fragmentIdsMap", async () => {
-      const piercer = new ReactPortablePiercer(fragmentIdMapping);
-      const rewriter = new HTMLRewriter().on(piercer.selector, piercer);
-
-      expect(await rewriter.transform(response).text()).matchSnapshot();
-    });
-  });
-
-  describe("FragmentTemplatesAppender", () => {
-    let response: Response;
-    beforeEach(() => {
-      response = new Response(hostDummyResponseBody);
-    });
-    test("empty fragmentIdsMap", async () => {
-      const appender = new FragmentTemplatesAppender(new Map());
-      const rewriter = new HTMLRewriter().on(appender.selector, appender);
-
-      expect(await rewriter.transform(response).text()).toBe(
-        hostDummyResponseBody,
-      );
-    });
-    test("fulfilled fragmentIdsMap", async () => {
-      const appender = new FragmentTemplatesAppender(fragmentIdMapping);
-      const rewriter = new HTMLRewriter().on(appender.selector, appender);
-
-      expect(await rewriter.transform(response).text()).matchSnapshot();
-    });
-  });
-
   describe("FragmentBaseReplacer", () => {
     describe("rp-fragment q:base", () => {
       let response: Response;
