+++ packages/gateway/src/tests/worker.ts
@@ -2,18 +2,14 @@ import { gateway } from "../index";
 
 export type Env = {
   ORIGIN: string;
-  FRAGMENT_CONFIGS: string;
-  ALLOW_ORIGINS: string;
-  CACHE: KVNamespace;
+  FRAGMENTS_ENDPOINT: string;
 };
 
 export default {
   fetch: (request: Request, env: Env, ctx: ExecutionContext) => {
     return gateway({
       proxy: env.ORIGIN,
-      cds: JSON.parse(env.FRAGMENT_CONFIGS),
-      cors: { origin: env.ALLOW_ORIGINS },
-      kv: env.CACHE,
+      fragmentsEndpoint: env.FRAGMENTS_ENDPOINT,
     })(request, env, ctx);
   },
 };
