+++ packages/gateway/vitest.config.ts
@@ -9,16 +9,7 @@ export default defineConfig({
       bindings: {
         ORIGIN: "https://origin.com",
         FRAGMENTS_ENDPOINT: "https://fragments.com",
-        ALLOW_ORIGINS: "*",
-        FRAGMENT_CONFIGS: JSON.stringify({
-          f1: { endpoint: "https://f1.com", assetPath: "" },
-          f2: {
-            endpoint: "https://f2.com",
-            assetPath: "https://assets.f2.com/statics",
-          },
-        }),
       },
-      kvNamespaces: ["TEST_KV", "CACHE"],
     },
   },
 });
